Split function tools from task_builder
This also fixes all mypy errors
